fix(work): extract shared cards, remove double-ellipsis, fix tab empty state (closes #134))
