+++ frontend/src/app/work/page.tsx
@@ -6,11 +6,11 @@ import { useProjects } from '@/lib/hooks/useProjects';
 import { usePublishedArticles } from '@/lib/hooks/useArticles';
 import { Project } from '@/lib/graphql/types/project.types';
 import { Article } from '@/lib/graphql/types/article.types';
-import { AlertCircle, Calendar, ExternalLink, Loader2 } from 'lucide-react';
+import { AlertCircle, ExternalLink, Loader2 } from 'lucide-react';
 import { Alert, AlertDescription } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
-import Link from 'next/link';
-import { formatDateSafe } from '@/utils/dateHandler';
+import { ProjectCard } from '@/components/work/ProjectCard';
+import { ArticleCard } from '@/components/work/ArticleCard';
 
 type Tab = 'all' | 'projects' | 'articles';
 
@@ -46,11 +46,6 @@ export default function WorkPage() {
         : allItems,
     [allItems, activeTab],
   );
-
-  const bothEmpty =
-    !loadingProjects && !loadingArticles &&
-    !projectsError && !articlesError &&
-    projects.length === 0 && articles.length === 0;
 
   return (
     <MainLayout>
@@ -126,10 +121,17 @@ export default function WorkPage() {
           </Alert>
         )}
 
-        {/* Global empty state — only when both sections have no data */}
-        {bothEmpty && visibleItems.length === 0 && (
+        {/* Empty state — tab-aware; suppressed during loading and when error banners already explain the gap */}
+        {!loading && visibleItems.length === 0 &&
+          !(activeTab === 'all' && (projectsError || articlesError)) && (
           <div className="text-center py-20">
-            <h3 className="text-lg font-semibold mb-2">Nothing here yet</h3>
+            <h3 className="text-lg font-semibold mb-2">
+              {activeTab === 'projects'
+                ? 'No projects yet'
+                : activeTab === 'articles'
+                ? 'No articles yet'
+                : 'Nothing here yet'}
+            </h3>
             <p className="text-muted-foreground">Content will appear once it is published.</p>
           </div>
         )}
@@ -148,134 +150,5 @@ export default function WorkPage() {
         )}
       </div>
     </MainLayout>
-  );
-}
-
-function ProjectCard({ project }: { project: Project }) {
-  return (
-    <Link href={`/projects/${project.slug}`} className="relative block group">
-      <div className="group rounded-lg border overflow-hidden bg-card text-card-foreground shadow hover:shadow-lg transition-all hover:scale-[1.02] h-full">
-        <div className="aspect-video w-full bg-muted relative overflow-hidden">
-          {project.imageUrl ? (
-            <div
-              className="absolute inset-0 bg-cover bg-center"
-              style={{ backgroundImage: `url(${project.imageUrl})` }}
-            />
-          ) : (
-            <div className="absolute inset-0 bg-gradient-to-br from-primary/20 to-background/80 flex items-center justify-center text-2xl font-bold">
-              {project.title}
-            </div>
-          )}
-        </div>
-
-        <div className="p-5 sm:p-7 flex flex-col gap-3 relative z-10">
-          <span className="text-[10px] font-semibold uppercase tracking-widest text-muted-foreground">
-            Project
-          </span>
-          <h2 className="text-xl font-semibold">{project.title}</h2>
-          <p className="text-muted-foreground text-sm line-clamp-3">
-            {project.description.length > 150
-              ? `${project.description.substring(0, 150)}...`
-              : project.description}
-          </p>
-
-          {project.technologies?.length > 0 && (
-            <div className="flex flex-wrap gap-2">
-              {project.technologies.map((tech, i) => (
-                <span
-                  key={i}
-                  className="bg-black text-gray-300 font-mono px-3 py-1.5 rounded-md text-xs border border-gray-800 shadow-[0_0_3px_#848884]"
-                >
-                  {tech}
-                </span>
-              ))}
-            </div>
-          )}
-
-          <div className="flex items-center text-sm text-muted-foreground">
-            <Calendar className="h-4 w-4 mr-1" />
-            <span>{formatDateSafe(project.createdAt)}</span>
-          </div>
-
-          <div className="flex gap-3">
-            {project.githubUrl && (
-              <span
-                role="link"
-                tabIndex={0}
-                onClick={(e) => {
-                  e.preventDefault();
-                  window.open(project.githubUrl, '_blank', 'noopener,noreferrer');
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter' || e.key === ' ')
-                    window.open(project.githubUrl, '_blank', 'noopener,noreferrer');
-                }}
-                className="text-sm font-medium hover:underline z-20 cursor-pointer"
-              >
-                GitHub
-              </span>
-            )}
-            <span className="text-sm font-medium text-primary hover:underline z-20">
-              View Details
-            </span>
-          </div>
-        </div>
-
-        <span className="absolute inset-0 z-10" aria-hidden="true" />
-      </div>
-    </Link>
-  );
-}
-
-function ArticleCard({ article }: { article: Article }) {
-  const preview =
-    article.excerpt ||
-    (article.content.length > 150 ? `${article.content.substring(0, 150)}...` : article.content);
-
-  return (
-    <Link href={`/articles/${article.slug}`} className="relative block group">
-      <div className="rounded-lg border overflow-hidden bg-card text-card-foreground shadow hover:shadow-lg transition-all hover:scale-[1.02] h-full">
-        {article.imageUrl && (
-          <div className="aspect-video w-full bg-muted overflow-hidden">
-            <div
-              className="w-full h-full bg-cover bg-center"
-              style={{ backgroundImage: `url(${article.imageUrl})` }}
-            />
-          </div>
-        )}
-
-        <div className="p-5 sm:p-7 flex flex-col gap-3 relative z-10">
-          <span className="text-[10px] font-semibold uppercase tracking-widest text-muted-foreground">
-            Article
-          </span>
-          <h2 className="text-xl font-semibold group-hover:text-primary transition-colors">
-            {article.title}
-          </h2>
-          <p className="text-muted-foreground text-sm line-clamp-3">{preview}</p>
-
-          {article.tags && article.tags.length > 0 && (
-            <div className="flex flex-wrap gap-2">
-              {article.tags.map((tag, i) => (
-                <span
-                  key={i}
-                  className="bg-secondary text-secondary-foreground px-2 py-1 rounded-full text-xs"
-                >
-                  {tag}
-                </span>
-              ))}
-            </div>
-          )}
-
-          <div className="flex items-center text-sm text-muted-foreground">
-            <Calendar className="h-4 w-4 mr-1" />
-            <span>{formatDateSafe(article.createdAt)}</span>
-          </div>
-
-          <span className="text-sm font-medium text-primary hover:underline z-20">Read more</span>
-        </div>
-
-        <span className="absolute inset-0 z-10" aria-hidden="true" />
-      </div>
-    </Link>
   );
 }
